fix: add team route and cleanup backend
