+++ app/models/index.js
@@ -6,8 +6,7 @@ import sequelize from "../config/sequelizeInstance.js";
 
 import User from "./user.model.js";
 import Session from "./session.model.js";
-import Tutorial from "./tutorial.model.js";
-import Lesson from "./lesson.model.js"; 
+import Team from "./team.model.js";
 
 
 const db = {};
@@ -16,8 +15,7 @@ db.sequelize = sequelize;
 
 db.user = User;
 db.session = Session;
-db.tutorial = Tutorial;
-db.lesson = Lesson;
+db.team = Team;
 
 // foreign key for session
 db.user.hasMany(
@@ -31,28 +29,7 @@ db.session.belongsTo(
   { foreignKey: { allowNull: false }, onDelete: "CASCADE" }
 );
 
-// foreign key for tutorials
-db.user.hasMany(
-  db.tutorial,
-  { as: "tutorial" },
-  { foreignKey: { allowNull: false }, onDelete: "CASCADE" }
-);
-db.tutorial.belongsTo(
-  db.user,
-  { as: "user" },
-  { foreignKey: { allowNull: false }, onDelete: "CASCADE" }
-);
-
-// foreign key for lessons
-db.tutorial.hasMany(
-  db.lesson,
-  { as: "lesson" },
-  { foreignKey: { allowNull: false }, onDelete: "CASCADE" }
-);
-db.lesson.belongsTo(
-  db.tutorial,
-  { as: "tutorial" },
-  { foreignKey: { allowNull: false }, onDelete: "CASCADE" }
-);
+// foreign key for teams
+// db.team.hasMany
 
 export default db;
