+++ app/routes/index.js
@@ -2,15 +2,13 @@ import { Router } from "express";
 
 import AuthRoutes from "./auth.routes.js";
 import UserRoutes from "./user.routes.js";
-import TutorialRoutes from "./tutorial.routes.js";
-import LessonRoutes from "./lesson.routes.js";
+import Team from "./team.routes.js";
 
 
 const router = Router();
 
 router.use("/", AuthRoutes);
 router.use("/users", UserRoutes);
-router.use("/tutorials", TutorialRoutes);
-router.use("/tutorials", LessonRoutes);
+router.use("/team", Team);
 
 export default router;
