--- app/routes/lesson.routes.js
@@ -1,32 +1,0 @@
-  import lessons from "../controllers/lesson.controller.js";
-  import authenticate from "../authorization/authorization.js";
-  import { Router } from "express";
-  var router = Router()
-
-  // Create a new Lesson for a Tutorial
-  router.post("/:tutorialId/lessons/", [authenticate], lessons.create);
-
-  // Retrieve all Lessons for a Tutorial
-  router.get(
-    "/:tutorialId/lessons/",
-    [authenticate],
-    lessons.findAllForTutorial
-  );
-
-  // Retrieve all published Lessons for a Tutorial
-  router.get(
-    "/:tutorialId/lessons/published",
-    [authenticate],
-    lessons.findAllPublished
-  );
-
-  // Retrieve a single Lesson with id
-  router.get("/:tutorialId/lessons/:id", [authenticate], lessons.findOne);
-
-  // Update a Lesson with id
-  router.put("/:tutorialId/lessons/:id", [authenticate], lessons.update);
-
-  // Delete a Lesson with id
-  router.delete("/:tutorialId/lessons/:id", [authenticate], lessons.delete);
-
-export default router
