+++ server.js
@@ -25,7 +25,6 @@ app.use(express.urlencoded({ extended: true }));
 // Load the routes from the routes folder
 app.use("/tracker-t3", routes); 
 
-
 // set port, listen for requests
 const PORT = process.env.PORT || 3100;
 if (process.env.NODE_ENV !== "test") {
